requirements.txt
anyio==3.5.0
asgiref==3.5.0
certifi==2021.10.8
chardet==3.0.4
click==7.1.2
fastapi==0.74.1
greenlet==1.1.2
gunicorn==20.0.4
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2021.12.1
httpcore==0.9.1
httpx==0.13.3
hyperframe==5.2.0
idna==2.10
pydantic==1.9.0
python-dotenv==0.19.2
rfc3986==1.5.0
sniffio==1.2.0
SQLAlchemy==1.4.31
starlette==0.17.1
typing_extensions==4.1.1
uvicorn==0.17.5
websockets==8.1
